[SPARK-14482][SQL] Change default compression codec for Parquet from gzip to snappy
